fix: refactor usagestats package
