+++ cmd/frontend/graphqlbackend/schema.go
@@ -2895,6 +2895,19 @@ enum UserEvent {
     CODEINTELREFS
     CODEINTELINTEGRATION
     CODEINTELINTEGRATIONREFS
+
+    # Product stages
+    STAGEMANAGE
+    STAGEPLAN
+    STAGECODE
+    STAGEREVIEW
+    STAGEVERIFY
+    STAGEPACKAGE
+    STAGEDEPLOY
+    STAGECONFIGURE
+    STAGEMONITOR
+    STAGESECURE
+    STAGEAUTOMATE
 }
 
 # A period of time in which a set of users have been active.
