+++ cmd/frontend/graphqlbackend/schema.graphql
@@ -2902,6 +2902,19 @@ enum UserEvent {
     CODEINTELREFS
     CODEINTELINTEGRATION
     CODEINTELINTEGRATIONREFS
+
+    # Product stages
+    STAGEMANAGE
+    STAGEPLAN
+    STAGECODE
+    STAGEREVIEW
+    STAGEVERIFY
+    STAGEPACKAGE
+    STAGEDEPLOY
+    STAGECONFIGURE
+    STAGEMONITOR
+    STAGESECURE
+    STAGEAUTOMATE
 }
 
 # A period of time in which a set of users have been active.
